requirements.txt
langchain==0.3.26
langchain-core==0.3.67
langchain-openai==0.3.27
langchain-google-genai==2.1.6
langgraph==0.5.0
aiohttp==3.12.13
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.1